Clean up projects in Dashboard
It's debatable whether I succeeded in "cleaning up" here. Added new 'projects' state to Dashboard to serve as single source of truth for projects. Now, rather than settinga 'currentProj' state that could diverge from the 'projects' field of 'currentUser', the 'projects' state serves as the definitive source of truth for project information. The currently selected project is controller by the 'currentProjectIndex' state, which should point to an element in the 'projects' state array. This does succeed in forcing SST, but make the code longer and more difficult to follow.